refactoring, working to add support for context menus
