+++ src/extension.ts
@@ -3,6 +3,7 @@ import * as vscode from 'vscode';
 import * as request from 'request-promise-native';
 import * as sqlops from 'sqlops';
 import { error } from 'util';
+import {placeScript} from './placescript';
 
 
 export function activate(context: vscode.ExtensionContext) {
@@ -16,7 +17,8 @@ export function activate(context: vscode.ExtensionContext) {
         };
         console.log('Bringing in the first responder kit from the mothership.');
         const scriptText = await request.get(options);
-        await placescript(fileName, scriptText);
+        //await placeScript.placescript(fileName, scriptText);
+        new placeScript().placescript(fileName,scriptText);
     };
     var disposable_spblitz = vscode.commands.registerCommand('extension.sp_blitz', getblitz);
     context.subscriptions.push(disposable_spblitz);
@@ -29,7 +31,8 @@ export function activate(context: vscode.ExtensionContext) {
         };
         console.log('Bringing in the first responder kit from the mothership.');
         const scriptText = await request.get(options);
-        await placescript(fileName, scriptText);
+        //await placeScript.placescript(fileName, scriptText);
+        new placeScript().placescript(fileName,scriptText);
     };
     var disposable_spblitzcache = vscode.commands.registerCommand('extension.sp_blitzcache', getblitzcache);
     context.subscriptions.push(disposable_spblitzcache);
@@ -42,7 +45,8 @@ export function activate(context: vscode.ExtensionContext) {
         };
         console.log('Bringing in the first responder kit from the mothership.');
         const scriptText = await request.get(options);
-        await placescript(fileName, scriptText);        
+        //await placeScript.placescript(fileName, scriptText);
+        new placeScript().placescript(fileName,scriptText);   
     };
     var disposable_spblitzfirst = vscode.commands.registerCommand('extension.sp_blitzfirst', getblitzfirst);
     context.subscriptions.push(disposable_spblitzfirst);
@@ -55,7 +59,8 @@ export function activate(context: vscode.ExtensionContext) {
         };
         console.log('Bringing in the first responder kit from the mothership.');
         const scriptText = await request.get(options);
-        await placescript(fileName, scriptText);
+        //await placeScript.placescript(fileName, scriptText);
+        new placeScript().placescript(fileName,scriptText);
     };
     var disposable_spblitzwho = vscode.commands.registerCommand('extension.sp_blitzwho', getblitzwho);
     context.subscriptions.push(disposable_spblitzwho);
@@ -74,7 +79,8 @@ export function activate(context: vscode.ExtensionContext) {
         -- uncomment the following line to write results to an output table
         --, @OutputDatabaseName = 'DBAtools', @OutputSchemaName = 'dbo', @OutputTableName = 'BlitzResults'`;
         var setting: vscode.Uri = vscode.Uri.parse("untitled:" + fileName);
-        await placescript(fileName, scriptText);
+        //await placeScript.placescript(fileName, scriptText);
+        new placeScript().placescript(fileName,scriptText);
     };
     var disposable_runspblitz = vscode.commands.registerCommand('extension.run_sp_blitz', runspblitz);
     context.subscriptions.push(disposable_runspblitz);
@@ -87,7 +93,8 @@ export function activate(context: vscode.ExtensionContext) {
         };
         console.log('Bringing in the first responder kit from the mothership.');
         const scriptText = await request.get(options);
-        await placescript(fileName, scriptText);
+        //await placeScript.placescript(fileName, scriptText);
+        new placeScript().placescript(fileName,scriptText);
     };
     var disposable_spblitzindex = vscode.commands.registerCommand('extension.sp_blitzindex', getblitzindex);
     context.subscriptions.push(disposable_spblitzindex);
@@ -101,7 +108,8 @@ export function activate(context: vscode.ExtensionContext) {
         --@TableName = '',
          @Mode = 4
          --0=Diagnose, 1=Summarize, 2=Index Usage Detail, 3=Missing Index Detail, 4=Diagnose Details`;
-        await placescript(fileName, scriptText);
+         //await placeScript.placescript(fileName, scriptText);
+         new placeScript().placescript(fileName,scriptText);
     };
     var disposable_runspblitzindex = vscode.commands.registerCommand('extension.run_sp_blitzindex', runspblitzindex);
     context.subscriptions.push(disposable_runspblitzindex);
@@ -115,7 +123,8 @@ export function activate(context: vscode.ExtensionContext) {
             -- CPU, executions, xpm, recent compilations, memory grant, writes, all
         @Top = 10
         `;
-        await placescript(fileName, scriptText);
+        //await placeScript.placescript(fileName, scriptText);
+        new placeScript().placescript(fileName,scriptText);
     };
     var disposable_runspblitzcache = vscode.commands.registerCommand('extension.run_sp_blitzcache',runspblitzcache)
     context.subscriptions.push(disposable_runspblitzcache);
@@ -129,7 +138,8 @@ export function activate(context: vscode.ExtensionContext) {
         @ShowSleepingSPIDs = 0,
         @ExpertMode = 0 --1 will also run sp_BlitzWho
         `;
-        await placescript(fileName, scriptText);
+        //await placeScript.placescript(fileName, scriptText);
+        new placeScript().placescript(fileName,scriptText);
     };
     var disposable_runspblitzfirst = vscode.commands.registerCommand('extension.run_sp_blitzfirst',runspblitzfirst);
     context.subscriptions.push(disposable_runspblitzfirst);
@@ -137,22 +147,22 @@ export function activate(context: vscode.ExtensionContext) {
 
 
     // places scriptText into fileName editor with current connection
-    async function placescript(fileName, scriptText) {
-        var setting: vscode.Uri = vscode.Uri.parse("untitled:" + fileName);
-        try {
-            let connection = await sqlops.connection.getCurrentConnection();
-            let doc = await vscode.workspace.openTextDocument(setting);
-            let editor = await vscode.window.showTextDocument(doc, 1, false);
-            editor.edit(edit => {
-                edit.insert(new vscode.Position(0, 0), scriptText);
-            });
-            if (connection) {
-                await sqlops.queryeditor.connect(doc.uri.toString(), connection.connectionId);
-            }
-        } catch (err) {
-            vscode.window.showErrorMessage(err);
-        }
-    }
+    // async function placescript(fileName, scriptText) {
+    //     var setting: vscode.Uri = vscode.Uri.parse("untitled:" + fileName);
+    //     try {
+    //         let connection = await sqlops.connection.getCurrentConnection();
+    //         let doc = await vscode.workspace.openTextDocument(setting);
+    //         let editor = await vscode.window.showTextDocument(doc, 1, false);
+    //         editor.edit(edit => {
+    //             edit.insert(new vscode.Position(0, 0), scriptText);
+    //         });
+    //         if (connection) {
+    //             await sqlops.queryeditor.connect(doc.uri.toString(), connection.connectionId);
+    //         }
+    //     } catch (err) {
+    //         vscode.window.showErrorMessage(err);
+    //     }
+    // }
 
 }
